+++ src/main/java/racingcar/domain/car/Cars.java
@@ -3,12 +3,11 @@ package racingcar.domain.car;
 import racingcar.domain.generator.OilGenerator;
 import racingcar.dto.CarPerformance;
 import racingcar.dto.RoundResult;
+import racingcar.exception.DuplicatedCarNameException;
 import racingcar.exception.NoCarsToFindMaxDistanceException;
 
 import java.util.Collections;
 import java.util.List;
-
-import static racingcar.constant.ExceptionMessage.DUPLICATED_CAR_NAME;
 
 public class Cars {
 
@@ -53,7 +52,7 @@ public class Cars {
                 .count();
 
         if (uniqueCarNamesCount != cars.size()) {
-            throw DUPLICATED_CAR_NAME.getException();
+            throw new DuplicatedCarNameException();
         }
     }
